Changed network to a Keras one, and created function to store and run previously trained networks
The neural network constructor now uses the Keras module. Genetic_algo.py still exists but is deprecated. GeneBis.py should be use instead to run the genetic algorithm training.
When the training is over, the best solution is stored in a .txt file. The script runExample.py can read into this file and run a game of Rogue with the weights of the previously stored solution.